Split up forge utilities and API doc template functions into separate modules: ``forge_utils`` and ``api_docs``
Also:
* Rename some functions and variables for consistency
* Remove template functions from Sphinx docs
