+++ pyinaturalist/api_docs.py
@@ -1,41 +1,9 @@
-# TODO: Consistent naming for template functions
 """
-Reusable template functions + utilities used for API documentation.
+Reusable template functions used for API documentation.
 Each template function contains a portion of an endpoint's request parameters, with corresponding
 type annotations and docstrings.
-
-Example usage::
-
-    # 1. Template function with individual request params + docs
-    >>> def get_foo_template(arg_1: str = None, arg_2: bool = False):
-    >>>     '''
-    >>>     arg_1: Example request parameter 1
-    >>>     arg_2: Example request parameter 2
-    >>>     '''
-
-    # 2. API Endpoint function with generic (variadic) keyword args
-    >>> @document_request_params(get_foo_template)
-    >>> def get_foo(**kwargs) -> List:
-        ''' Get Foo resource '''
-
-    # 3. Modified function signature + docstring
-    >>> help(get_foo)
-    '''
-    Help on function get_foo:
-
-    get_foo(arg_1: str = None, arg_2: bool = False) -> List
-      Get Foo resource
-
-      Args:
-          arg_1: Example request parameter 1
-          arg_2: Example request parameter 2
-    '''
 """
-from inspect import cleandoc
-from itertools import chain
-from functools import wraps
-from logging import getLogger
-from typing import Any, Callable, Dict, List
+from typing import Any, Dict, List
 
 from pyinaturalist.constants import (
     MultiInt,
@@ -43,110 +11,12 @@ from pyinaturalist.constants import (
     Date,
     DateTime,
     IntOrStr,
-    TemplateFunction,
 )
 from pyinaturalist.request_params import MULTIPLE_CHOICE_PARAMS
 
-logger = getLogger(__name__)
-
-
-def document_request_params(template_functions: List[TemplateFunction]):
-    """Document a function with both docstrings and function signatures from one or more
-    template functions.
-
-    Signature modification requires ``forge``. If not installed, only docstrings will be modified.
-
-    Args:
-        template_functions: Template functions containing docstrings and params to apply to the
-            wrapped function
-    """
-    template_functions += [user_agent]
-
-    def f(func):
-        # Modify docstring
-        func = copy_docstrings(func, template_functions)
-
-        # If forge is installed, modify signature; otherwise, silently ignore it
-        try:
-            func = copy_signatures(func, template_functions)
-        except ImportError:
-            logger.debug("Forge not installed; skipping runtime API documentation")
-
-        # Call modified function
-        @wraps(func)
-        def g(*args, **kwargs):
-            return func(*args, **kwargs)
-
-        return g
-
-    return f
-
-
-def copy_docstrings(
-    target_function: Callable, template_functions: List[TemplateFunction]
-) -> Callable:
-    """Copy docstrings from one or more template functions to a target function.
-    Assumes Google-style docstrings.
-
-    Args:
-        target_function: Function to modify
-        template_functions: Functions containing docstrings to apply to ``target_function``
-    """
-    # Start with initial function description only
-    docstring, return_value_desc = _split_docstring(target_function.__doc__ or "")
-
-    # Insert 'Args' section
-    docstring += "\n\nArgs:"
-    for template_function in template_functions:
-        docstring += template_function.__doc__ or ""
-        docstring = docstring.rstrip()
-
-    # Insert 'Returns' section, if present
-    if return_value_desc:
-        docstring += "\n\nReturns:\n    " + return_value_desc
-
-    target_function.__doc__ = docstring
-    return target_function
-
-
-def _split_docstring(docstring):
-    """ Split a docstring into a function description + return value description, if present. """
-    if "Returns:" in docstring:
-        function_desc, return_value_desc = docstring.split("Returns:")
-    else:
-        function_desc = docstring
-        return_value_desc = ""
-
-    return cleandoc(function_desc.strip()), cleandoc(return_value_desc.strip())
-
-
-def copy_signatures(
-    target_function: Callable, template_functions: List[TemplateFunction]
-) -> Callable:
-    """Copy function signatures from one or more template functions to a target function.
-
-    Args:
-        target_function: Function to modify
-        template_functions: Functions containing params to apply to ``target_function``
-    """
-    revision = _get_combined_revision(template_functions)
-    return revision(target_function)
-
-
-def _get_combined_revision(template_functions: List[TemplateFunction]):
-    """ Create a :py:class:`forge.Revision` from the combined parameters of multiple functions """
-    import forge
-
-    # Use forge.copy to create a revision for each template function
-    revisions = [forge.copy(func) for func in template_functions]
-
-    # Combine the parameters of all revisions into a single revision
-    fparams = [list(rev.signature.parameters.values()) for rev in revisions]
-    return forge.sign(*list(chain.from_iterable(fparams)))
-
 
 # Params that are in most observation-related endpoints in both Node and REST APIs
-def observation_params_common(
+def _observation_params_common(
     q: str = None,
     d1: Date = None,
     d2: Date = None,
@@ -183,7 +53,7 @@ def observation_params_common(
 
 
 # Observation params that are only in the Node API
-def observation_params_node_only(
+def _observation_params_node_only(
     acc: bool = None,
     captive: bool = None,
     endemic: bool = None,
@@ -323,7 +193,7 @@ def observation_params_node_only(
 
 
 # Observation params that are only in the REST API
-def observation_params_rest_only(
+def _observation_params_rest_only(
     has: MultiStr = None,
     on: Date = None,
     m1: Date = None,
@@ -411,7 +281,7 @@ def _update_observation_params(
     """
 
 
-def taxon_params(
+def _taxon_params(
     q: str = None,
     is_active: bool = None,
     taxon_id: int = None,
@@ -441,7 +311,7 @@ def taxon_params(
     """
 
 
-def taxon_id_params(
+def _taxon_id_params(
     id_above: int = None,
     id_below: int = None,
     only_id: int = None,
@@ -459,14 +329,14 @@ def taxon_id_params(
 # ------------------------
 
 
-def access_token(access_token: str = None):
+def _access_token(access_token: str = None):
     """
     access_token: An access token required for user authentication, as returned by
         :py:func:`get_access_token()`
     """
 
 
-def bounding_box(
+def _bounding_box(
     nelat: float = None,
     nelng: float = None,
     swlat: float = None,
@@ -480,56 +350,57 @@ def bounding_box(
     """
 
 
-def geojson_properties(properties: List[str] = None):
+def _geojson_properties(properties: List[str] = None):
     """
     properties: Properties from observation results to include as GeoJSON properties
     """
 
 
-def legacy_params(params: Dict[str, Any] = None):
+# TODO: Remove deprecated `params` arg in 0.12
+def _legacy_params(params: Dict[str, Any] = None):
     """
     params: [DEPRECATED] Request parameters as a dict instead of keyword arguments
     """
 
 
-def minify(minify: str = None):
+def _minify(minify: str = None):
     """
     minify: Condense each match into a single string containg taxon ID, rank, and name
     """
 
 
-def name(name: str = None):
+def _name(name: str = None):
     """
     name: Name must match this value
     """
 
 
-def only_id(only_id: bool = False):
+def _only_id(only_id: bool = False):
     """
     only_id: Return only the record IDs
     """
 
 
-def observation_id(observation_id: int):
+def _observation_id(observation_id: int):
     """
     observation_id: iNaturalist observation ID
     """
 
 
-def page(page: int = None):
+def _page(page: int = None):
     """
     page: Page number of results to return
     """
 
 
-def pagination(
+def _pagination(
     page: int = None,
     per_page: int = None,
     order: str = None,
     order_by: str = None,
 ):
     """
-    page: Pagination page number
+    page: Page number of results to return
     per_page: Number of results to return in a page. The maximum value is generally 200,
         unless otherwise noted
     order: Sort order
@@ -538,16 +409,9 @@ def pagination(
 
 
 # TODO: Remove deprecated `search_query` kwarg in 0.12
-def search_query(q: str = None, search_query: str = None):
+def _search_query(q: str = None, search_query: str = None):
     """
     q: Search query
-    """
-
-
-# Added to every function signature by default
-def user_agent(user_agent: str = None):
-    """
-    user_agent: A custom user-agent string to provide to the iNaturalist API
     """
 
 
@@ -557,39 +421,46 @@ def _format_param_choices():
     )
 
 
+MULTIPLE_CHOICE_PARAM_DOCS = "**Multiple-Choice Parameters:**\n" + _format_param_choices()
+
+
 # Request param combinations for Node API endpoints
+# Note: user_agent param is added to all functions by default
+# ------------------------------------------------------------
+
 _get_observations = [
-    legacy_params,
-    observation_params_common,
-    observation_params_node_only,
-    bounding_box,
+    _legacy_params,
+    _observation_params_common,
+    _observation_params_node_only,
+    _bounding_box,
 ]
-get_observations_params = _get_observations + [pagination, only_id]
-get_all_observations_params = _get_observations + [only_id]
+
+get_observations_params = _get_observations + [_pagination, _only_id]
+get_all_observations_params = _get_observations + [_only_id]
 get_observation_species_counts_params = _get_observations
-get_geojson_observations_params = _get_observations + [geojson_properties]
-get_places_nearby_params = [bounding_box, name]
-get_taxa_params = [taxon_params, taxon_id_params]
-get_taxa_autocomplete_params = [taxon_params, minify]
+get_geojson_observations_params = _get_observations + [_geojson_properties]
+get_places_nearby_params = [_bounding_box, _name]
+get_taxa_params = [_taxon_params, _taxon_id_params]
+get_taxa_autocomplete_params = [_taxon_params, _minify]
+
 
 # Request param combinations for REST API endpoints
+# ------------------------------------------------------------
+
 get_observations_params_rest = [
-    observation_params_common,
-    observation_params_rest_only,
-    bounding_box,
-    pagination,
+    _observation_params_common,
+    _observation_params_rest_only,
+    _bounding_box,
+    _pagination,
 ]
-get_observation_fields_params = [search_query, page]
-get_all_observation_fields_params = [search_query]
-create_observations_params = [legacy_params, access_token, _create_observations_params]
+get_observation_fields_params = [_search_query, _page]
+get_all_observation_fields_params = [_search_query]
+create_observations_params = [_legacy_params, _access_token, _create_observations_params]
 update_observation_params = [
-    observation_id,
-    legacy_params,
-    access_token,
+    _observation_id,
+    _legacy_params,
+    _access_token,
     _create_observations_params,
     _update_observation_params,
 ]
-delete_observation_params = [observation_id, access_token]
-
-
-MULTIPLE_CHOICE_PARAM_DOCS = "**Multiple-Choice Parameters:**\n" + _format_param_choices()
+delete_observation_params = [_observation_id, _access_token]
